+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release 2020.5 (?)
 
+## Bug fixes
+- Fixed import loop when using `inmanta.execute.proxy` as entry point (#2341)
+
 # Release 2020.4 (2020-09-08)
 
 ## New features